Update script
Updated the script to use "du" instead of "stat" for checking if the filesystem has not been written to via the Sync Module

The previous check using "stat" would always pass stability checks due to "stat" passing the logical size of the backing files, not the physical size.  Since the backing file is emulating a block media, this would cause the sectors on the emulated media to become corrupt due to the script unbinding while the Sync Module was still writing to the file.